# Conflicts:
#	src/main/java/com/digibank/restapi/model/entity/User.java

+++ src/main/java/com/digibank/restapi/service/PasswordService.java
@@ -4,7 +4,7 @@ import com.digibank.restapi.dto.changePassword.ChangePasswordDto;
 import com.digibank.restapi.dto.createPassword.CreatePasswordDto;
 
 public interface PasswordService {
-    CreatePasswordDto changePassword(Integer id_user, CreatePasswordDto request);
+    CreatePasswordDto changePassword(Long id_user, CreatePasswordDto request);
 
-    CreatePasswordDto changePasswordWithValidation(Integer id_user, ChangePasswordDto changePasswordDto);
+    CreatePasswordDto changePasswordWithValidation(Long id_user, ChangePasswordDto changePasswordDto);
 }
